+++ apps/testapp-embedded/src/app.d.ts
@@ -11,10 +11,11 @@ declare global {
 }
 
 declare module 'virtual:sdocs' {
-	import type { AxisConfig, DocEntry } from 'sdocs';
+	import type { AxisConfig, DocEntry, ScaleConfig } from 'sdocs';
 	export const docs: DocEntry[];
 	export const cssNames: string[];
 	export const axes: Required<AxisConfig>[];
+	export const scale: Required<ScaleConfig> | null;
 	export const pageModules: Record<string, () => Promise<{ default: unknown }>>;
 	export default docs;
 }